requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.32
psycopg[binary]==3.2.10
python-dotenv==1.0.1
pymongo==4.8.0
